Change routes for getting tasks by project and creating task by project
